完成说说配置
